% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqFromWaves.R
\name{getColumnIndex}
\alias{getColumnIndex}
\title{Column indexes of protoname in the table returned by extractSeqFromWaves}
\usage{
getColumnIndex(data, protoname)
}
\arguments{
\item{data}{Tibble or data.frame. Table returned by \code{\link{seqFromWaves}}.}

\item{protoname}{String. Protoname of a variable.}
}
\value{
Vector of indexes of the columns of \code{data} corresponding to the protoname.
}
\description{
Column indexes in \code{data} of the variable fitting the provided protoname.
}
\details{
Returns the column indexes in \code{data} of the variable fitting the provided protoname.
Examples in help page of \code{\link{seqFromWaves}}.
}
\seealso{
\code{\link{seqFromWaves}}
}
